requirements.txt
flask==2.1.2
flask_sqlalchemy==2.5.1
flask_cors==3.0.10
flask_restful==0.3.9
flask_migrate==3.1.0
flask_wtf==1.0.1
flask-caching==2.0.0
dependency-injector==4.39.1
pyyaml==6.0
requests==2.28.1
gunicorn==20.1.0
gevent==21.12.0
eventlet==0.33.1
celery==5.2.7
pymysql==1.0.2
numpy==1.23.1
blinker==1.5
PyJWT==2.4.0
webargs==8.2.0
passlib==1.7.4
redis==4.3.4